Move the code for adding code completion definitions to Monaco instance from the sql editor page to a hook.
